+++ AppleAuth/AppleAuthQuickLoginArgs.cs
@@ -1,0 +1,12 @@
+namespace AppleAuth
+{
+    public struct AppleAuthQuickLoginArgs
+    {
+        public readonly string Nonce;
+
+        public AppleAuthQuickLoginArgs(string nonce = null)
+        {
+            this.Nonce = nonce;
+        }
+    }
+}